Whitelisting.py: simplify verification
The JSON for whitelistings is organized by package. This changes the
code to make use of that: it only reads the audits for the currently
tested package and verifies files against that.

This gets rid of some complex nested loops checking file paths against
package names while trying to avoid complaining about the same things
repeatedly.